try code refactoring..fails example
